Changes the internal resolution implementation (The scene can now be drawn to an off-screen buffer texture, which can be blit to the screen. This texture has the same dimensions as the internal resolution)

+++ include/Rect.h
@@ -17,10 +17,10 @@ public:
     inline int w() const { return _w; }
     inline int h() const { return _h; }
 	
-	inline int left() { return x(); }
-	inline int right() { return x() + _w; } 
-	inline int top() { return y(); }
-	inline int bottom() { return y() + _h; } 
+	inline float left() { return x(); }
+	inline float right() { return x() + _w; } 
+	inline float top() { return y(); }
+	inline float bottom() { return y() + _h; } 
 
     void setPosition(Vector2D& position) {
         this->_position.setX(position.x());
